+++ internal/ceres/block_random_access_matrix.h
@@ -119,8 +119,8 @@ class CERES_EXPORT_INTERNAL BlockRandomAccessMatrix {
 
   // Number of scalar rows and columns in the matrix, i.e the sum of
   // all row blocks and column block sizes respectively.
-  virtual int num_rows() const = 0;
-  virtual int num_cols() const = 0;
+  virtual int64_t num_rows() const = 0;
+  virtual int64_t num_cols() const = 0;
 };
 
 }  // namespace internal
