+++ internal/ceres/sparse_matrix.h
@@ -98,9 +98,9 @@ class CERES_EXPORT_INTERNAL SparseMatrix : public LinearOperator {
   virtual double* mutable_values() = 0;
   virtual const double* values() const = 0;
 
-  virtual int num_rows() const = 0;
-  virtual int num_cols() const = 0;
-  virtual int num_nonzeros() const = 0;
+  virtual int64_t num_rows() const = 0;
+  virtual int64_t num_cols() const = 0;
+  virtual int64_t num_nonzeros() const = 0;
 };
 
 }  // namespace internal
